Rename setd/kdebase-workspace-4.11 -> sets/kdebase-workspace-4.11_init6
Strangely that file was not there in my kde overlay clone. It should
have been there, according to the log, so I have no idea what was
causing that stray. Maybe the file was out for lunch?
After updating the kde overlay it showed up again, so this should be
renamed in the init6 overlay, too.